DISCO-3172: Consolidate test fixtures
