+++ tests/integration/conftest.py
@@ -4,7 +4,6 @@
 pytest_plugins = [
     "tests.integration.fixtures.gcs",
     "tests.integration.fixtures.metrics",
-    "tests.integration.fixtures.app_client",
     "tests.integration.fixtures.image_manifest",
     "tests.integration.api.v1.curated_recommendations.corpus_backends.fixtures",
 ]
